backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.4.2
langchain==0.0.300
sentence-transformers==2.2.2
pymongo==4.5.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
pdfplumber==0.10.2
pytesseract==0.3.10
loguru==0.7.0
python-dotenv==1.0.0
groq==0.4.2
spacy==3.7.2
regex==2023.10.3
transformers==4.35.0
asyncio==3.4.3
motor==3.3.2
aiofiles==23.2.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
